toml file is now consistent with dir structure
The tutorial was explicit about the directory structure including head-wasm32-unknown-wasi-full but did not mention renaming and moving of files so as to make it work with the toml file as listed. To help a first time reader of toml understand its various properties, we update it to match the directory structure as described earlier in the tutorial.

Signed-off-by: John-Mason P. Shackelford <jpshack@gmail.com>